Ensure `ivy-tab-panel`'s `isVisible` property is a Boolean.
Previously it was an alias to `active`, which returns a String so that
it can be used as a classname binding. The problem is, Ember observes
`isVisible` and calls jQuery's `toggle` function while passing in the
value of `isVisible` to determine whether to show or hide the element.
In this case, `isVisible` was returning a String, which jQuery would
interpret as an easing function. This would result in the tab panels
animating in and out when switching between tabs.