Remove some pesky `tagName` assumptions
Previously, `ivy-tab-list` defaulted to a `<ul>` element, and `ivy-tab`
defaulted to an `<li>` element. This commit removes the default
`tagName` from `ivy-tab-list` and makes `ivy-tab` default to an `<a>`
tag, which has slightly better semantic meaning for a tab. As well, this
commit adds an `href` attribute to `ivy-tab`, which points to the
`ivy-tab-panel` it controls. If `tagName` is set to something other than
'a', then the `href` attribute will not be set.