fix: address linting errors
This addresses errors when running "yarn lint" which involved
applying a codemod to change all the `this.get(` patterns into
`getless` versions.  Additionally there were two computed properties
that were not declaring dependencies on the dependent computed property.

A linting error remains with the use of an improper role `presentation` on
an `li`.