fix: remove role presentation from li, update readme
The specification of the presentation role on the `li` tags creates
a linting error within the new template linting in ember. The role
seems to have no functional impact on screenreaders, so it is
removed.

The example in the readme was updated to agree with the examples in
the dummy application.