feat: add ability to override default a-tag href
This adds a new property to the tag component, `fixedHref`, that when overridden
will allow the href of an `a` tag version of the tag component to
be specified to a fixed url.  This allows the tabs to be bookmarked,
as well as used to open new browser windows.

A new test page is added to demonstrate tabs as subroutes that
leverage the `fixedHref`.